requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
npm==0.1.1
optional-django==0.1.0
psycopg2==2.9.5
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.5.0
Werkzeug==2.2.2
